Progressive Web App (PWA)

+++ client/src/components/ad-request/PendingAdRequestList.vue
@@ -36,6 +36,11 @@ async function acceptAdRequest(adRequestId) {
 async function rejectAdRequest(adRequestId) {
     console.log('Ad Request rejected!');
     try {
+        const confirmReject = window.confirm("Are you sure you want to reject this ad request?");
+        if (!confirmReject) {
+            return null;
+        }
+
         const res = await fetch(`/api/ad-request/${adRequestId}/reject`, {
             method: 'GET',
             headers: { 'Authentication-Token': localStorage.getItem('authToken') }
@@ -76,6 +81,11 @@ async function getPendingAdRequests() {
 async function deleteAdRequest(adRequestId) {
     console.log('Ad Request Deleted!');
     try {
+        const confirmDelete = window.confirm("Are you sure you want to delete this ad request?");
+        if (!confirmDelete) {
+            return null;
+        }
+        
         const res = await fetch(`/api/ad-request/${adRequestId}`, {
             method: 'DELETE',
             headers: { 'Authentication-Token': localStorage.getItem('authToken') }
@@ -130,7 +140,7 @@ onMounted(async () => {
 
                             <div class="d-flex gap-2 mt-auto">
                                 <RouterLink 
-                                    :to="`/ad-request/${ad.id}/edit`" 
+                                    :to="`/ad-request/${ad.id}/edit?influencer_id=${ad.influencer_id}`" 
                                     class="btn btn-warning btn-sm"
                                 >
                                     Edit
